use correct allocation size.
`vsizeof` only adds space for the Object header, but none of the Cown data
afterwards. Therefor we need to pass in the `CownData<T>` instead of just `T`.

C++ makes this confusing by using `T` sometimes for the usertype (eg `i32`,
and sometimes as the self type in CRTP). Maybe I should send something upstream.